+++ tests/test_coverage_sorteddict.py
@@ -215,6 +215,11 @@ def test_values():
     temp = SortedDict(mapping)
     assert list(temp.values()) == [pos for key, pos in mapping]
 
+def test_iterkeys():
+    temp = SortedDict()
+    with pytest.raises(AttributeError):
+        temp.iterkeys
+
 def test_notgiven():
     assert repr(SortedDict._SortedDict__not_given) == '<not-given>'